+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -3,7 +3,7 @@ package roomescape.config;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.controller.MemberArgumentResolver;
+import roomescape.controller.security.MemberArgumentResolver;
 import roomescape.service.AuthService;
 
 import java.util.List;
